Fixed signed/unsigned comparison

+++ src/SFML/Graphics/Font.cpp
@@ -46,7 +46,7 @@ namespace
     unsigned long Read(FT_Stream rec, unsigned long offset, unsigned char* buffer, unsigned long count)
     {
         sf::InputStream* stream = static_cast<sf::InputStream*>(rec->descriptor.pointer);
-        if (stream->Seek(offset) == offset)
+        if (static_cast<unsigned long>(stream->Seek(offset)) == offset)
         {
             if (count > 0)
                 return static_cast<unsigned long>(stream->Read(reinterpret_cast<char*>(buffer), count));
